ADSR envelope
Created an ADSR envelope (class AdsrEnvelope). The monophonic
synthesizer now uses that to amplify the oscillator accordingly.

See issue 6 regarding frequency discontinuities for this.

Also added a lowpass filter for the oscillator frequency (with rather
high cutoff frequency, i.e. the frequency changes are still very fast
and not really audible as a glissando).

Added a lowpass filter in the monophonic synthesizer that removes
audible clicks which occured with an attack or release phase of zero.

Created a new class AudioSource, which serve as base class for all
classes that generate audio (or control signals) at a given sample rate.