Record to memory and waveform view
Integrated the GraphView class from an old project, which can show
QAbstractTableModel instances that carry numeric data.

Also integrated a corresponding model, FloatTableModel, and a subclass
JackAudioModel, which is capable to read audio from a Jack ring buffer.

Together with the new class Record2MemoryClient it is now possible to
trigger recording on an input audio port via a MIDI note trigger on a
MIDI input port and show the result in a widget.

Fixed an error regarding missing member initialization in the Oscillator
class. Also improved the PiecewiseLinearOscillator class to be capable
of generating jitter-free sound at any frequency.

Fixed an error in AdsrEnvelope.

Modified the main window to incorporate an audio view and start a
synthesizer and a record client at startup to test the new classes.