2-pole Butterwirth filter
Based on the general z-plane filter. Given a normalized cutoff frequency
the coefficients are computed. From these, the poles and zeros are
derived. Then the coefficients are once again calculated, but now from
the poles and zeros.

Note: this should be changed to just computing the coefficients from the
cutoff frequency and forgetting about the poles and zeros, i.e., not
extending ZPlaneFilter.