Base widget for controlling Jack clients
Implemented a base class for GUIs controlling Jack clients via MIDI. The
concept is the following:
- given a arbitrary Jack client with MIDI input (currently this has to
be a JackClient object, but in theory any Jack client with a MIDI input
would do it)
- a widget derives from MidiClientWidget
- the MidiClientWidget class internally creates a Midi2SignalClient
object and connects its MIDI output to the MIDI input of the given Jack
client
- the widget communicates with the Midi2SignalClient object via signals
and slots
- the Midi2SignalClient object communicates with the given Jack client
solely via MIDI

I.e., GUI events have to be translated to MIDI events and vice versa.

The Midi2SignalClient class also has a MIDI input which can be used to
reflect changes from outside (via MIDI) in the GUI. For this the Jack
clients connected to this MIDI input should also be connected to the
given Jack client's MIDI input.

Possible improvement: change Midi2SignalClient such that it is able to
forward MIDI from its input to its output. This would spare the
additional connection mentioned above. Note that the MIDI events from
input and GUI have to be merged correctly with regard to their time!

Also added a test widget class TestMidiClientWidget, which just has two
buttons which activate and deactivate both internal clients.