Changed JackClient to report the actual client name if changed by Jack
The name given to the constructor will be handed to Jack. If that name
is already taken by a client, Jack will modify it to make it unique.
While the client is active, getName() will return the actual (possibly
modified) name. Is the client is inactive, getName() will return the
requested name.