Keyboard graphics item with mouse interaction and signals/slots
Implemented a GraphicsKeyboardItem class and a corresponding
GraphicsKeyItem class representing a piano keyboard and its keys,
respectively. Both are subclasses of QGraphicsItem.

GraphicsKeyboardItem is also a QObject and has signals and slots which
enable mouse interaction (pressing and releasing keys) with the
keyboard.

Modified MainWindow to show the possibilities of the new classes:
pressing a key on the keyboard will change the filter's cutoff frequency
to the frequency of the note associated with the pressed key.