IIR filter frequency inversion and Moog filter zeros
Added a method to IIRFilter which allows to invert the frequency
response of a filter by changing the coefficients. Inverting here means
to switch high and low frequencies. Essentially this is just inverting
the sign of all zeros and poles, which corresponds to changing the sign
of all odd coefficients.

Modified the IIRButterworthFilter to allow for selection between filter
types low-pass and high-pass. The low-pass filter is realized using the
new IIRFilter::invert() method (see above).

Added the possibility to specifiy the number of zeros at DC (for a low-
filters IIRButterworthFilter and IIRMoogFilter. Default is two zeros for
the IIRButterworthFilter (which also has 2 poles) and no zeros for the
IIRMoogFilter.