Parallel and serial connection of IIR filters
Implemented an addition and a multiplication operator in IIRFilter.
Adding two IIR filters essentially means operating them in parallel
(summing the output of two independent filters). Multiplying two IIR
filters means operating them in serial (filtering the output of one
filter with the other).

While this is a costly operation involving multiplications of
polynomials, and thus might not be useful for realtime, it makes it
easier to look at the resulting frequency spectrum.

Modified MainWindow to show the parallel operation of a low-pass and a
high-pass Butterworth filter, as well as the serial operation of two
low-pass Butterworth filters (resulting in a 4-pole low-pass filter).