Midi input output via signals&slots II
Reimplemented the functionality of Midi2SignalClient with the new
classes JackClientWithDeferredProcessing and JackThread. The
functionality is now split over two classes: MidiClient and MidiThread.
Both communicate over the new class JackRingBuffer<T>.

The class Midi2SignalClient is now obsolete and will be removed in a
future commit (when the new classes have been tested).