Made IirMoogFilter process MIDI note input
The base cutoff frequency is set to the input note's frequency.

Also changed GraphicsNodeItem to include reverse scaling. As a result,
the node can be positioned in the graphics view by giving the scaled
position (and not only the graphics view based position).

Changed MainWindow to show the changes. By pressing a key on the virtual
keyboard, the frequency response view changes accordingly. The realtime
Moog filter will also change its base frequency if its MIDI input is
connected to the "midi signal client" MIDI output.