Adjusted waveshaping function to  map from and to [-1..1]
Also made it possible to have neighbouring control points with the same
x coordinate in LinearWaveshapingClient (changed
LinearInterpolatorIntegral accordingly).

Changed the notification parameters in LinearWaveshapingClient such that
it is now possible to have varying numbers of control points.

Added a method to EventProcessor that allows for sending multiple event
at the same time (and of course with the same time stamp).