Implemented dynamic port creation for the meta_jack wrapper client
At startup the wrapper client has no ports. The internal interface
clients however (one for input and one for output) have one port for
audio and one for midi. Whenever an unconnected port is connected to an
internal client, a new unconnected port of the same type is created
internally, and the connected port is made visible to the JACK server by
creating an appropriate port on the wrapper client.

Not yet implemented is the removal of disconnected ports from the
wrapper client.