Created new generic ring buffer based on QObject
and a corresponding EventProcessorClient2 class.

With these, Jack clients are not restricted to one type of event coming
from outside the Jack process thread.

This is due to the following scenario: the Oscillator class should have
a gain event which can be used to set the loudness of the oscillator
from the GUI. Subclasses of Oscillator however have their own event
types for setting certain oscillator properties.