Fixed some bugs in metajack
Still not quite clear:

- can a client register ports while not activating? In the current
implementation it can... and the ports can even be connected. But that
shouldn't matter in the current implementation. Their process callback
won't be run until they're activated.
- when will port unregister events be sent if ports are not explicitly
unregistered? In the current implementation they are sent when the
client is closed, after it has been activated, but before the client
unregister event is sent.