Implementation of session load/save for most clients
Most clients now can save and load their state to a QDataStream.

This is used by MainWindow to make it possible save/load the current
session to/from a given file, including which clients are active and
their positions and the connections between them.

One bug: OscillatorGraphicsItem does not show the correct gain after a
load. The gain of the underying oscillator however seems to be correct.