Removed port shapes from the client shape
(having the client shape overlay the ports disables the mouse events
sent to the ports).

One issue remaining: When a port receives a mouse event, the client item
loses focus (despite the port being one of its children) and thus may be
sent to the back of the graphics scene (covered by other client items).

Maybe reimplement the mouse event handlers in the client item to send
the events to the corresponding port child while still being focused?