New method in JackContext to get a client handle by name
The implementation in RealJackContext always returns 0, as there is no
way in the original Jack implementation to get a client handle by the
client name.

The MetaJackContext implementation however returns the client handle, if
a client with that name exists in that context.

Purpose of the method is to be able to associate client names with the
corresponding JackClient instance (via the corresponding jack_client_t
pointer).