Rolled back the two last changes to timing and adapted all classes accordingly
This means that all time information handed to methods like
processAudio(), processMidi(), processNoteOn(), ..., and processEvent()
now receive a time parameter in the range of [0,nframes).

Classes that need unique time values that increase monotonically and
linearly even exceeding the bounds of a process cycle need to take care
of that themselves.